python-backend/requirements.txt
fastapi
mangum
openai
python-dotenv